backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
coreapi==2.3.3
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.74
tensorflow==2.13.0
ultralytics==8.0.196
easyocr==1.7.0
pyyaml==6.0.1
uritemplate==4.1.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django-filter==23.5
django-guardian==2.4.0
django-storages==1.14.2
django-cleanup==8.0.0
django-rest-auth==0.9.5
django-allauth==0.57.0
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.5.0
dj-database-url==2.1.0
python-magic==0.4.27
requests==2.31.0
PyJWT==2.8.0
